save conda envs to cache
