+++ hippunfold/plugins/atlas.py
@@ -16,19 +16,10 @@ logger = logging.getLogger(__name__)
 
 import json
 import os
+from hippunfold.workflow.lib import utils as utils
 
 # Global variable to store the commit hash
 ATLAS_REPO_COMMIT = "c1a53ecade939ead9de8f9169c6a4ddff0c73c3d"
-
-
-def get_download_dir():
-    if "HIPPUNFOLD_CACHE_DIR" in os.environ.keys():
-        download_dir = os.environ["HIPPUNFOLD_CACHE_DIR"]
-    else:
-        # create local download dir if it doesn't exist
-        dirs = AppDirs("hippunfold", "khanlab")
-        download_dir = dirs.user_cache_dir
-    return download_dir
 
 
 def sync_atlas_repo():
@@ -36,7 +27,7 @@ def sync_atlas_repo():
     Ensures the atlas folder is synced from the public GitHub repository using GitPython.
     """
     repo_url = "https://github.com/khanlab/hippunfold-atlases.git"
-    atlas_dir = Path(get_download_dir()) / "hippunfold-atlases"
+    atlas_dir = Path(utils.get_download_dir()) / "hippunfold-atlases"
 
     try:
         if atlas_dir.exists() and (atlas_dir / ".git").exists():
@@ -98,7 +89,7 @@ def get_atlas_configs():
     sync_atlas_repo()
 
     # Define search locations
-    cache_dir = get_download_dir()
+    cache_dir = utils.get_download_dir()
 
     atlas_dirs = []
     atlas_dirs.append(Path(cache_dir) / "hippunfold-atlases")
