+++ hippunfold/workflow/rules/shape_inject.smk
@@ -178,11 +178,7 @@ rule template_shape_reg:
         conda_env("greedy")
     threads: 8
     log:
-        bids_log_wrapper(
-            "template_shape_reg",
-            **inputs.subj_wildcards,
-            hemi="{hemi}"
-        ),
+        bids_log_wrapper("template_shape_reg", **inputs.subj_wildcards, hemi="{hemi}"),
     shell:
         #affine (with moments), then greedy
         "greedy -threads {threads} {params.general_opts} {params.affine_opts} {params.img_pairs} -o {output.matrix}  &> {log} && "
